+++ Godeps/_workspace/src/github.com/onsi/ginkgo/ginkgo/run_and_watch_command_flags.go
@@ -2,8 +2,9 @@ package main
 
 import (
 	"flag"
-	"github.com/onsi/ginkgo/config"
 	"runtime"
+
+	"github.com/onsi/ginkgo/config"
 )
 
 type RunAndWatchCommandFlags struct {
@@ -14,6 +15,8 @@ type RunAndWatchCommandFlags struct {
 	Cover          bool
 	Notify         bool
 	SkipPackage    string
+	Tags           string
+	AutoNodes      bool
 
 	//not for watch command
 	KeepGoing    bool
@@ -38,6 +41,31 @@ func NewWatchCommandFlags(flagSet *flag.FlagSet) *RunAndWatchCommandFlags {
 	return c
 }
 
+func (c *RunAndWatchCommandFlags) wasSet(flagName string) bool {
+	wasSet := false
+	c.FlagSet.Visit(func(f *flag.Flag) {
+		if f.Name == flagName {
+			wasSet = true
+		}
+	})
+
+	return wasSet
+}
+
+func (c *RunAndWatchCommandFlags) computeNodes() {
+	if c.wasSet("nodes") {
+		return
+	}
+	if c.AutoNodes {
+		switch n := runtime.NumCPU(); {
+		case n <= 4:
+			c.NumCPU = n
+		default:
+			c.NumCPU = n - 1
+		}
+	}
+}
+
 func (c *RunAndWatchCommandFlags) flags(forWatchCommand bool) {
 	onWindows := (runtime.GOOS == "windows")
 	onOSX := (runtime.GOOS == "darwin")
@@ -45,6 +73,7 @@ func (c *RunAndWatchCommandFlags) flags(forWatchCommand bool) {
 	config.Flags(c.FlagSet, "", false)
 
 	c.FlagSet.IntVar(&(c.NumCPU), "nodes", 1, "The number of parallel test nodes to run")
+	c.FlagSet.BoolVar(&(c.AutoNodes), "p", false, "Run in parallel with auto-detected number of nodes")
 	c.FlagSet.BoolVar(&(c.ParallelStream), "stream", onWindows, "stream parallel test output in real time: less coherent, but useful for debugging")
 	c.FlagSet.BoolVar(&(c.Recurse), "r", false, "Find and run test suites under the current directory recursively")
 	c.FlagSet.BoolVar(&(c.Race), "race", false, "Run tests with race detection enabled")
@@ -53,6 +82,7 @@ func (c *RunAndWatchCommandFlags) flags(forWatchCommand bool) {
 	if onOSX {
 		c.FlagSet.BoolVar(&(c.Notify), "notify", false, "Send desktop notifications when a test run completes")
 	}
+	c.FlagSet.StringVar(&(c.Tags), "tags", "", "A list of build tags to consider satisfied during the build")
 	if !forWatchCommand {
 		c.FlagSet.BoolVar(&(c.KeepGoing), "keepGoing", false, "When true, failures from earlier test suites do not prevent later test suites from running")
 		c.FlagSet.BoolVar(&(c.UntilItFails), "untilItFails", false, "When true, Ginkgo will keep rerunning tests until a failure occurs")