+++ Godeps/_workspace/src/github.com/onsi/ginkgo/ginkgo/watch_command.go
@@ -3,6 +3,9 @@ package main
 import (
 	"flag"
 	"fmt"
+	"time"
+
+	"github.com/onsi/ginkgo/config"
 	"github.com/onsi/ginkgo/ginkgo/testrunner"
 	"github.com/onsi/ginkgo/ginkgo/testsuite"
 )
@@ -38,6 +41,7 @@ type SpecWatcher struct {
 }
 
 func (w *SpecWatcher) WatchSpecs(args []string, additionalArgs []string) {
+	w.commandFlags.computeNodes()
 	w.notifier.VerifyNotificationsAreAvailable()
 
 	suites := findSuites(args, w.commandFlags.Recurse, w.commandFlags.SkipPackage)
@@ -68,12 +72,22 @@ func (w *SpecWatcher) WatchSuites(suites []*testsuite.TestSuite, additionalArgs 
 }
 
 func (w *SpecWatcher) RunSuite(suite *testsuite.TestSuite, additionalArgs []string) {
-	runner := testrunner.New(suite, w.commandFlags.NumCPU, w.commandFlags.ParallelStream, w.commandFlags.Race, w.commandFlags.Cover, additionalArgs)
+	w.UpdateSeed()
+	runner := testrunner.New(suite, w.commandFlags.NumCPU, w.commandFlags.ParallelStream, w.commandFlags.Race, w.commandFlags.Cover, w.commandFlags.Tags, additionalArgs)
 	err := runner.Compile()
 	if err != nil {
 		fmt.Print(err.Error())
 	}
-	suitePassed := (err == nil) && runner.Run()
-	w.notifier.SendSuiteCompletionNotification(suite, suitePassed)
+	runResult := testrunner.FailingRunResult()
+	if err == nil {
+		runResult = runner.Run()
+	}
+	w.notifier.SendSuiteCompletionNotification(suite, runResult.Passed)
 	runner.CleanUp()
+}
+
+func (w *SpecWatcher) UpdateSeed() {
+	if !w.commandFlags.wasSet("seed") {
+		config.GinkgoConfig.RandomSeed = time.Now().Unix()
+	}
 }