+++ cf/i18n/init.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/XenoPhex/jibber_jabber"
+	"github.com/pivotal-cf-experimental/jibber_jabber"
 
 	resources "github.com/cloudfoundry/cli/cf/resources"
 	go_i18n "github.com/nicksnyder/go-i18n/i18n"